refactor(net): introduce BaseSignalingClient and implement authoritative lobby actions
